+++ tests/e2e/core-tests/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ## Added
 
+- Shopper Checkout Login Account
 - Shopper My Account Create Account
 
 ## Fixed